+++ tests/audio.rs
@@ -61,6 +61,12 @@ fn test_audio_interchange_file_format() {
 }
 
 #[test]
+fn test_creative_voice() {
+    let format = FileFormat::from_file("fixtures/audio/sample.voc").unwrap();
+    assert_eq!(format, FileFormat::CreativeVoice);
+}
+
+#[test]
 fn test_fast_tracker2_extended_module() {
     let format = FileFormat::from_file("fixtures/audio/sample.xm").unwrap();
     assert_eq!(format, FileFormat::FastTracker2ExtendedModule);
@@ -85,9 +91,21 @@ fn test_monkeys_audio() {
 }
 
 #[test]
-fn test_mpeg_audio_layer3() {
+fn test_mpeg12_audio_layer3() {
     let format = FileFormat::from_file("fixtures/audio/sample.mp3").unwrap();
-    assert_eq!(format, FileFormat::MpegAudioLayer3);
+    assert_eq!(format, FileFormat::Mpeg12AudioLayer3);
+}
+
+#[test]
+fn test_mpeg1_audio_layer1() {
+    let format = FileFormat::from_file("fixtures/audio/sample.mp1").unwrap();
+    assert_eq!(format, FileFormat::Mpeg1AudioLayer1);
+}
+
+#[test]
+fn test_mpeg1_audio_layer2() {
+    let format = FileFormat::from_file("fixtures/audio/sample.mp2").unwrap();
+    assert_eq!(format, FileFormat::Mpeg1AudioLayer2);
 }
 
 #[test]
@@ -133,9 +151,9 @@ fn test_qualcomm_pure_voice() {
 }
 
 #[test]
-fn test_scream_tracker3_module() {
+fn test_screamtracker3_module() {
     let format = FileFormat::from_file("fixtures/audio/sample.s3m").unwrap();
-    assert_eq!(format, FileFormat::ScreamTracker3Module);
+    assert_eq!(format, FileFormat::Screamtracker3Module);
 }
 
 #[test]
@@ -145,13 +163,13 @@ fn test_sony_dsd_stream_file() {
 }
 
 #[test]
-fn test_wav_pack() {
-    let format = FileFormat::from_file("fixtures/audio/sample.wv").unwrap();
-    assert_eq!(format, FileFormat::WavPack);
-}
-
-#[test]
 fn test_waveform_audio() {
     let format = FileFormat::from_file("fixtures/audio/sample.wav").unwrap();
     assert_eq!(format, FileFormat::WaveformAudio);
+}
+
+#[test]
+fn test_wavpack() {
+    let format = FileFormat::from_file("fixtures/audio/sample.wv").unwrap();
+    assert_eq!(format, FileFormat::Wavpack);
 }